Move MySql and SqlLite execution tests into subdir.
