+++ QueryBuilder.Tests/MySql/MySqlExecutionTest.cs
@@ -1,30 +1,26 @@
-using SqlKata.Compilers;
-using Xunit;
-using SqlKata.Execution;
-using MySql.Data.MySqlClient;
 using System;
-using System.Linq;
-using static SqlKata.Expressions;
 using System.Collections.Generic;
-using Microsoft.Data.Sqlite;
+using System.Linq;
+using Microsoft.Extensions.Configuration;
+using MySql.Data.MySqlClient;
+using SqlKata.Compilers;
+using SqlKata.Execution;
+using Xunit;
 
-namespace SqlKata.Tests
+namespace SqlKata.Tests.MySql
 {
-    public class SqliteExecutionTest
+    public class MySqlExecutionTest
     {
         [Fact]
         public void EmptySelect()
         {
 
             var db = DB().Create("Cars", new[] {
-                    "Id INTEGER PRIMARY KEY AUTOINCREMENT",
+                    "Id INT PRIMARY KEY AUTO_INCREMENT",
                     "Brand TEXT NOT NULL",
                     "Year INT NOT NULL",
                     "Color TEXT NULL",
             });
-
-
-            var tables = db.Select(@"SELECT name FROM sqlite_schema WHERE type ='table' AND name NOT LIKE 'sqlite_%'");
 
             var rows = db.Query("Cars").Get();
 
@@ -37,7 +33,7 @@ namespace SqlKata.Tests
         public void SelectWithLimit()
         {
             var db = DB().Create("Cars", new[] {
-                    "Id INTEGER PRIMARY KEY AUTOINCREMENT",
+                    "Id INT PRIMARY KEY AUTO_INCREMENT",
                     "Brand TEXT NOT NULL",
                     "Year INT NOT NULL",
                     "Color TEXT NULL",
@@ -53,49 +49,10 @@ namespace SqlKata.Tests
         }
 
         [Fact]
-        public void InsertGetId()
-        {
-            var db = DB().Create("Cars", new[] {
-                    "Id INTEGER PRIMARY KEY AUTOINCREMENT",
-                    "Brand TEXT NOT NULL",
-                    "Year INT NOT NULL",
-            });
-
-            var id = db.Query("Cars").InsertGetId<int>(new
-            {
-                Brand = "Toyota",
-                Year = 1900
-            });
-
-            Assert.Equal(1, id);
-
-            id = db.Query("Cars").InsertGetId<int>(new
-            {
-                Brand = "Toyota 2",
-                Year = 1901
-            });
-
-            Assert.Equal(2, id);
-
-            id = db.Query("Cars").InsertGetId<int>(new
-            {
-                Brand = "Toyota 2",
-                Year = 1901
-            });
-
-            Assert.Equal(3, id);
-
-
-            db.Drop("Cars");
-        }
-
-
-
-        [Fact]
         public void Count()
         {
             var db = DB().Create("Cars", new[] {
-                    "Id INTEGER PRIMARY KEY AUTOINCREMENT",
+                    "Id INT PRIMARY KEY AUTO_INCREMENT",
                     "Brand TEXT NOT NULL",
                     "Year INT NOT NULL",
                     "Color TEXT NULL",
@@ -122,7 +79,7 @@ namespace SqlKata.Tests
         public void CloneThenCount()
         {
             var db = DB().Create("Cars", new[] {
-                    "Id INTEGER PRIMARY KEY AUTOINCREMENT",
+                    "Id INT PRIMARY KEY AUTO_INCREMENT",
                     "Brand TEXT NOT NULL",
                     "Year INT NOT NULL",
                     "Color TEXT NULL",
@@ -151,7 +108,7 @@ namespace SqlKata.Tests
         public void QueryWithVariable()
         {
             var db = DB().Create("Cars", new[] {
-                    "Id INTEGER PRIMARY KEY AUTOINCREMENT",
+                    "Id INT PRIMARY KEY AUTO_INCREMENT",
                     "Brand TEXT NOT NULL",
                     "Year INT NOT NULL",
                     "Color TEXT NULL",
@@ -181,7 +138,7 @@ namespace SqlKata.Tests
         public void InlineTable()
         {
             var db = DB().Create("Transaction", new[] {
-                    "Id INTEGER PRIMARY KEY AUTOINCREMENT",
+                    "Id INT PRIMARY KEY AUTO_INCREMENT",
                     "Amount int NOT NULL",
                     "Date DATE NOT NULL",
             });
@@ -198,7 +155,7 @@ namespace SqlKata.Tests
                     new object[] {"2022-01-01", 0.5},
                 })
                 .Join("Rates", "Rates.Date", "Transaction.Date")
-                .SelectRaw("([Transaction].[Amount] * [Rates].[Rate]) as AmountConverted")
+                .SelectRaw("Transaction.Amount * Rates.Rate as AmountConverted")
                 .Get();
 
             Assert.Single(rows);
@@ -207,13 +164,47 @@ namespace SqlKata.Tests
             db.Drop("Transaction");
         }
 
+        [Fact]
+        public void ExistsShouldReturnFalseForEmptyTable()
+        {
+            var db = DB().Create("Transaction", new[] {
+                    "Id INT PRIMARY KEY AUTO_INCREMENT",
+                    "Amount int NOT NULL",
+                    "Date DATE NOT NULL",
+            });
 
+            var exists = db.Query("Transaction").Exists();
+            Assert.Equal(false, exists);
+
+            db.Drop("Transaction");
+        }
+
+        [Fact]
+        public void ExistsShouldReturnTrueForNonEmptyTable()
+        {
+            var db = DB().Create("Transaction", new[] {
+                    "Id INT PRIMARY KEY AUTO_INCREMENT",
+                    "Amount int NOT NULL",
+                    "Date DATE NOT NULL",
+            });
+
+            db.Query("Transaction").Insert(new
+            {
+                Date = "2022-01-01",
+                Amount = 10
+            });
+
+            var exists = db.Query("Transaction").Exists();
+            Assert.Equal(true, exists);
+
+            db.Drop("Transaction");
+        }
 
         [Fact]
         public void BasicSelectFilter()
         {
             var db = DB().Create("Transaction", new[] {
-                    "Id INTEGER PRIMARY KEY AUTOINCREMENT",
+                    "Id INT PRIMARY KEY AUTO_INCREMENT",
                     "Date DATE NOT NULL",
                     "Amount int NOT NULL",
             });
@@ -222,12 +213,12 @@ namespace SqlKata.Tests
                 // 2020
                 {"2020-01-01", 10},
                 {"2020-05-01", 20},
-                
+
                 // 2021
                 {"2021-01-01", 40},
                 {"2021-02-01", 10},
                 {"2021-04-01", -10},
-                
+
                 // 2022
                 {"2022-01-01", 80},
                 {"2022-02-01", -30},
@@ -258,13 +249,21 @@ namespace SqlKata.Tests
             db.Drop("Transaction");
         }
 
+        private static string GetConnectionString()
+        {
+            var config = new ConfigurationBuilder()
+                .AddUserSecrets(typeof(MySqlExecutionTest).Assembly)
+                .Build();
+            return config.GetConnectionString("MySqlExecutionTest");
+        }
+
+        private static Lazy<string> ConnectionString { get; } = new(GetConnectionString);
+
         QueryFactory DB()
         {
-            var cs = $"Data Source=file::memory:;Cache=Shared";
+            var connection = new MySqlConnection(ConnectionString.Value);
 
-            var connection = new SqliteConnection(cs);
-
-            var db = new QueryFactory(connection, new SqliteCompiler());
+            var db = new QueryFactory(connection, new MySqlCompiler());
 
             return db;
         }
