+++ QueryBuilder.Tests/Sqlite/SQLiteExecutionTest.cs
@@ -1,27 +1,27 @@
-using SqlKata.Compilers;
-using Xunit;
-using SqlKata.Execution;
-using MySql.Data.MySqlClient;
-using System;
-using System.Linq;
-using static SqlKata.Expressions;
 using System.Collections.Generic;
-using Microsoft.Extensions.Configuration;
+using System.Linq;
+using Microsoft.Data.Sqlite;
+using SqlKata.Compilers;
+using SqlKata.Execution;
+using Xunit;
 
-namespace SqlKata.Tests
+namespace SqlKata.Tests.Sqlite
 {
-    public class MySqlExecutionTest
+    public class SqliteExecutionTest
     {
         [Fact]
         public void EmptySelect()
         {
 
             var db = DB().Create("Cars", new[] {
-                    "Id INT PRIMARY KEY AUTO_INCREMENT",
+                    "Id INTEGER PRIMARY KEY AUTOINCREMENT",
                     "Brand TEXT NOT NULL",
                     "Year INT NOT NULL",
                     "Color TEXT NULL",
             });
+
+
+            var tables = db.Select(@"SELECT name FROM sqlite_schema WHERE type ='table' AND name NOT LIKE 'sqlite_%'");
 
             var rows = db.Query("Cars").Get();
 
@@ -34,7 +34,7 @@ namespace SqlKata.Tests
         public void SelectWithLimit()
         {
             var db = DB().Create("Cars", new[] {
-                    "Id INT PRIMARY KEY AUTO_INCREMENT",
+                    "Id INTEGER PRIMARY KEY AUTOINCREMENT",
                     "Brand TEXT NOT NULL",
                     "Year INT NOT NULL",
                     "Color TEXT NULL",
@@ -50,10 +50,49 @@ namespace SqlKata.Tests
         }
 
         [Fact]
+        public void InsertGetId()
+        {
+            var db = DB().Create("Cars", new[] {
+                    "Id INTEGER PRIMARY KEY AUTOINCREMENT",
+                    "Brand TEXT NOT NULL",
+                    "Year INT NOT NULL",
+            });
+
+            var id = db.Query("Cars").InsertGetId<int>(new
+            {
+                Brand = "Toyota",
+                Year = 1900
+            });
+
+            Assert.Equal(1, id);
+
+            id = db.Query("Cars").InsertGetId<int>(new
+            {
+                Brand = "Toyota 2",
+                Year = 1901
+            });
+
+            Assert.Equal(2, id);
+
+            id = db.Query("Cars").InsertGetId<int>(new
+            {
+                Brand = "Toyota 2",
+                Year = 1901
+            });
+
+            Assert.Equal(3, id);
+
+
+            db.Drop("Cars");
+        }
+
+
+
+        [Fact]
         public void Count()
         {
             var db = DB().Create("Cars", new[] {
-                    "Id INT PRIMARY KEY AUTO_INCREMENT",
+                    "Id INTEGER PRIMARY KEY AUTOINCREMENT",
                     "Brand TEXT NOT NULL",
                     "Year INT NOT NULL",
                     "Color TEXT NULL",
@@ -80,7 +119,7 @@ namespace SqlKata.Tests
         public void CloneThenCount()
         {
             var db = DB().Create("Cars", new[] {
-                    "Id INT PRIMARY KEY AUTO_INCREMENT",
+                    "Id INTEGER PRIMARY KEY AUTOINCREMENT",
                     "Brand TEXT NOT NULL",
                     "Year INT NOT NULL",
                     "Color TEXT NULL",
@@ -109,7 +148,7 @@ namespace SqlKata.Tests
         public void QueryWithVariable()
         {
             var db = DB().Create("Cars", new[] {
-                    "Id INT PRIMARY KEY AUTO_INCREMENT",
+                    "Id INTEGER PRIMARY KEY AUTOINCREMENT",
                     "Brand TEXT NOT NULL",
                     "Year INT NOT NULL",
                     "Color TEXT NULL",
@@ -139,7 +178,7 @@ namespace SqlKata.Tests
         public void InlineTable()
         {
             var db = DB().Create("Transaction", new[] {
-                    "Id INT PRIMARY KEY AUTO_INCREMENT",
+                    "Id INTEGER PRIMARY KEY AUTOINCREMENT",
                     "Amount int NOT NULL",
                     "Date DATE NOT NULL",
             });
@@ -156,7 +195,7 @@ namespace SqlKata.Tests
                     new object[] {"2022-01-01", 0.5},
                 })
                 .Join("Rates", "Rates.Date", "Transaction.Date")
-                .SelectRaw("Transaction.Amount * Rates.Rate as AmountConverted")
+                .SelectRaw("([Transaction].[Amount] * [Rates].[Rate]) as AmountConverted")
                 .Get();
 
             Assert.Single(rows);
@@ -165,47 +204,13 @@ namespace SqlKata.Tests
             db.Drop("Transaction");
         }
 
-        [Fact]
-        public void ExistsShouldReturnFalseForEmptyTable()
-        {
-            var db = DB().Create("Transaction", new[] {
-                    "Id INT PRIMARY KEY AUTO_INCREMENT",
-                    "Amount int NOT NULL",
-                    "Date DATE NOT NULL",
-            });
 
-            var exists = db.Query("Transaction").Exists();
-            Assert.Equal(false, exists);
-
-            db.Drop("Transaction");
-        }
-
-        [Fact]
-        public void ExistsShouldReturnTrueForNonEmptyTable()
-        {
-            var db = DB().Create("Transaction", new[] {
-                    "Id INT PRIMARY KEY AUTO_INCREMENT",
-                    "Amount int NOT NULL",
-                    "Date DATE NOT NULL",
-            });
-
-            db.Query("Transaction").Insert(new
-            {
-                Date = "2022-01-01",
-                Amount = 10
-            });
-
-            var exists = db.Query("Transaction").Exists();
-            Assert.Equal(true, exists);
-
-            db.Drop("Transaction");
-        }
 
         [Fact]
         public void BasicSelectFilter()
         {
             var db = DB().Create("Transaction", new[] {
-                    "Id INT PRIMARY KEY AUTO_INCREMENT",
+                    "Id INTEGER PRIMARY KEY AUTOINCREMENT",
                     "Date DATE NOT NULL",
                     "Amount int NOT NULL",
             });
@@ -250,21 +255,13 @@ namespace SqlKata.Tests
             db.Drop("Transaction");
         }
 
-        private static string GetConnectionString()
-        {
-            var config = new ConfigurationBuilder()
-                .AddUserSecrets(typeof(MySqlExecutionTest).Assembly)
-                .Build();
-            return config.GetConnectionString("MySqlExecutionTest");
-        }
-
-        private static Lazy<string> ConnectionString { get; } = new(GetConnectionString);
-
         QueryFactory DB()
         {
-            var connection = new MySqlConnection(ConnectionString.Value);
+            var cs = $"Data Source=file::memory:;Cache=Shared";
 
-            var db = new QueryFactory(connection, new MySqlCompiler());
+            var connection = new SqliteConnection(cs);
+
+            var db = new QueryFactory(connection, new SqliteCompiler());
 
             return db;
         }
